Move response decode to adapter
Since req handles response body decoding for us, move the
hackney-specific response handling code into the adapter module.
