+++ test/honeybadger/http_adapter/hackney_test.exs
@@ -123,4 +123,63 @@ defmodule Honeybadger.HTTPAdapter.HackneyTest do
                ])
     end
   end
+
+  @body %{"a" => "1", "b" => "2"}
+  @headers [{"content-type", "application/json"}]
+  @json_library (Code.ensure_loaded?(JSON) && JSON) || Jason
+  @json_encoded_body @json_library.encode!(@body)
+  @uri_encoded_body URI.encode_query(@body)
+
+  test "decode_response_body/2" do
+    assert {:ok, response} =
+             Hackney.decode_response_body(
+               %HTTPResponse{body: @json_encoded_body, headers: @headers},
+               []
+             )
+
+    assert response.body == @body
+
+    assert {:ok, response} =
+             Hackney.decode_response_body(
+               %HTTPResponse{
+                 body: @json_encoded_body,
+                 headers: [{"content-type", "application/json; charset=utf-8"}]
+               },
+               []
+             )
+
+    assert response.body == @body
+
+    assert {:ok, response} =
+             Hackney.decode_response_body(
+               %HTTPResponse{
+                 body: @json_encoded_body,
+                 headers: [{"content-type", "text/javascript"}]
+               },
+               []
+             )
+
+    assert response.body == @body
+
+    assert {:ok, response} =
+             Hackney.decode_response_body(
+               %HTTPResponse{
+                 body: @uri_encoded_body,
+                 headers: [{"content-type", "application/x-www-form-urlencoded"}]
+               },
+               []
+             )
+
+    assert response.body == @body
+
+    assert {:ok, response} =
+             Hackney.decode_response_body(%HTTPResponse{body: @body, headers: []}, [])
+
+    assert response.body == @body
+
+    assert {:error, %Honeybadger.InvalidResponseError{} = error} =
+             Hackney.decode_response_body(%HTTPResponse{body: "%", headers: @headers}, [])
+
+    assert error.response.body == "%"
+  end
 end
